# Handle lack of RAW and WAR atomic ordering in message-order-fence