--- a/test/runtime/configMatters/comm/atomic-ordering.good
+++ b/test/runtime/configMatters/comm/atomic-ordering.good
@@ -1,0 +1,1 @@
+success