… test/runtime/configMatters/comm/atomic-ordering.numlocales