Change the message-order-fence implementation to not require read-after-write (`FI_ORDER_ATOMIC_RAW`) and write-after-read (`FI_ORDER_ATOMIC_WAR`) ordering for atomic operations. As an initial fix the `ofi` comm layer now:

* doesn't ask for these orderings when looking for a message-order-fence provider

* fences all fetching atomic operations. This ensures read-after-write ordering, while write-after-read is ensured by the MCM. All fetching atomic operations are fenced, eventually it should be done selectively only when a fetching atomic follows a non-fetching.

See https://github.com/Cray/chapel-private/issues/4140 for details.

Resolves https://github.com/Cray/chapel-private/issues/4140 and closes https://github.com/Cray/chapel-private/issues/4254.

Also resolves https://github.com/Cray/chapel-private/issues/4138.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>