@ronawho 2023-01-26
Can you check the performance impact of this? My initial thought would be to run comm-ops with 1, 8, 64, and 128 threads. Could also updated ra-atomic to use fetching AMOs and see the perf impact on something a little bigger than the microbenchmark.