@jhh67 2023-01-26
Performance results running on an HPE Cray EX:

**comm-ops**

% chpl comm-ops.chpl --fast --no-cache-remote --no-optimize-forall-unordered-ops -scommDiagsPrintUnstable

% env CHPL_RT_NUM_THREADS_PER_LOCALE=1 ./comm-ops -nl 2

| Benchmark | main  | PR    |
|-----------|-------|-------|
| GET       | 1.63s | 1.47s |
| PUT       | 0.13s | 0.11s |
| FAMO      | 1.72s | 1.95s |
| NFAMO     | 0.28s | 0.49s |

% env CHPL_RT_NUM_THREADS_PER_LOCALE=8 ./comm-ops -nl 2

| Benchmark | main  | PR    |
|-----------|-------|-------|
| GET       | 1.77s | 1.77s |
| PUT       | 0.96s | 0.91s |
| FAMO      | 1.95s | 2.64s |
| NFAMO     | 3.28s | 2.73s |

% env CHPL_RT_NUM_THREADS_PER_LOCALE=64 ./comm-ops -nl 2

| Benchmark | main   | PR     |
|-----------|--------|--------|
| GET       | 1.72s  | 1.85s  |
| PUT       | 1.40s  | 4.47s  |
| FAMO      | 1.78s  | 2.66s  |
| NFAMO     | 13.69s | 20.21s |

% env CHPL_RT_NUM_THREADS_PER_LOCALE=128 ./comm-ops -nl 2

| Benchmark | main   | PR     |
|-----------|--------|--------|
| GET       | 1.80s  | 1.82s  |
| PUT       | 9.35s  | 9.25s  |
| FAMO      | 1.85s  | 2.78s  |
| NFAMO     | 40.74s | 27.35s |

**ra-atomics**

These are the performance results of running `ra-atomics` with the following modification:
```
-    T(r & indexMask).xor(r);
+    T(r & indexMask).fetchXor(r);
```
% start_test --perflabel ml- test/release/examples/benchmarks/hpcc/ra-atomics.chpl

| main          | PR            |
|---------------|---------------|
| 0.440083 GUPS | 0.315457 GUPS |