@ronawho 2023-01-26
In that table `FAMO` is really the only thing we're expecting to have changed, so the `PUT`/`NFAMO` seems like expected noise from known thread-hot issues.  Let's get rid of the PUT/NFAMO rows since I think the noise just makes the results confusing. 

Can you recheck the `GET` results? I was expecting those to be unchanged and fairly stable.