@jhh67 2023-01-26
I removed both the PUT and NFAMO rows from the table. I also re-ran the benchmarks and entered the new results. The GET  results still jump around.