@ronawho 2023-01-26
Ok, let's remove GET too and just keep this about FAMO