@jhh67 2023-01-26
I used `salloc` to use the same two nodes for all `comm-ops` tests and now the results are more stable. Should we still remove GET?