@ronawho 2023-01-27
Can you add the reproducer from https://github.com/Cray/chapel-private/issues/4140 as a test in `test/runtime/configMatters/comm/`?

I have also confirmed that this resolves https://github.com/Cray/chapel-private/issues/4138 (please add to PR description)