@jhh67 2023-01-27
Added the test case and updated the PR description.